configure: fix check for bash shell
The bash shell sets the BASH variable with the full path name used to
execute the current instance of bash. But the configure script uses it
to attempt to check if the bash binary is present on a given system.

The configure script uses the #!/bin/sh shebang, so on a system where
/bin/sh is the bash shell, the configure script will fail to find the
bash binary due BASH being set to /bin/sh:

checking for bash... /bin/sh
configure: WARNING: Required executable bash not found, system tests require a bash shell!

Use instead a variable name that isn't an internal one used by bash.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>