tpm2_create: support pipe based stdin
The file routines used for reading from stdin uses
ftell() under the hood to get the file size and then
veirfy that sizes are sane. In the case of stdin
occuring from a pipe, an Illegal Seek is trigered
from ftell().

Correct this by supressing error reporting and attempting
a fall back file read mechanism.

Fixes: #425

Signed-off-by: William Roberts <william.c.roberts@intel.com>