lib: sort TPM2_RC entries for base response code values
The response code can come from different layers of the software stack,
and these are defined in the entries of the tpm2_tss_layer_entry array.

Sort the entries using the base value for each response code group. It
makes easier to map the entries to the response code levels defined in
the specifications and add more entries.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>