tpm2_rc_decode: add support for Resource Manager level errors
Add entries for the TSS2_RESMGR_ERROR_LEVEL and TSS2_RESMGRTPM_ERROR_LEVEL
error levels in the TPM2_RC array. This will allow the tpm2_rc_decode tool
to decode response codes sent by the Resource Manager layer.

For example, if a resource manager validates the command header instead of
leaving it to the TPM2 and returns a command response buffer with the code
for command not supported, the response code would be decoded as following:

$ tpm2_rc_decode 0xb0143
error layer
  hex: 0xb0000
  identifier: TSS2_RESMGRTPM_ERROR_LEVEL
  description: Error from the Resource Manager duplicating TPM error check
format 0 error code
  hex: 0x43
  name: TPM2_RC_COMMAND_CODE
  description: command code not supported

Fixes: #647

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>