options: remove unused ':' case option and error messages
By default the getopt_long() function prints an error message and returns
a '?' character to indicate an error. And it only returns a ':' on error
if the optstring starts with a colon.

So the ':' switch case wasn't used and also the printed error is unneeded
since getopt_long() already prints this.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>