configure: get version from a file
Use the idea in tpm2-abrmd to generate a VERSION file and use this file
to get the software version instead of relying on the fact we're in an
active git tree (this is not the case when we fetch a tarball).

The VERSION file generated by bootstrap is then included in the
distribution.