tpm2_clearlock: add tool
The tpm2_clearlock tools is used to authorize/forbid the use of
tpm2_clear. It used TPM2_ClearControl() to set or clear disableClear
(when disableClear is set it's not possible to call TPM2_Clear()).

One can check for the status of disableClear using the tpm2_getcap tool.

The clearlock.sh test check for the behavior of this newly added tool.

Fixes: #710

Signed-off-by: Emmanuel Deloget <logout@free.fr>