command line: change option names from output to out-file
When comparing command line switches that are specified to select an
output file, most of the tools use option 'out-file' and only a few of
them use option 'output' or other rarer names such as 'outfile'. In
order to rationalize option naming a bit, we turn all these option
names into 'out-file'.

Fixes: #729

Signed-off-by: Emmanuel Deloget <logout@free.fr>