refactor session and policy designs
Split out the policy building and session initiation code.

The session code has a lot of inputs, use a builder pattern
to build a data structure with sane session defaults and
allow the caller to override if needed.

By splitting the policy building code out from session
creation, this reduces the number of arguments required
to the policy routines.

This also (finally), documents the internal APIs for session
and policy handling, and adds much needed unit tests for
session handling. Tests for policy building will be
done in subsequent patches.

Signed-off-by: William Roberts <william.c.roberts@intel.com>