lib: use the exelp() exec* wrapper instead of execvpe()
The execvpe() wrapper is a GNU extension, but there isn't a reason to use
it. The command-line arguments array only contains the man pager command
and page, and the calling program's environment variables is used anyways.

So just use the execlp() wrapper that has the same behaviour, but it's a
POSIX standard execve() wrapper and leads simpler code.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>