tests/encryptdecrypt: fix errors that cause passing
The error trap was cleared when using tpm2_getcap
to verify if command encryptdecrypt was supported.

If it wasn't supported, by checking the grep return
code via $?, then a warning is printed and the test
is skipped.

When tpm2_getcap fails, $? is 1 and the test is reported
as passing, which is incorrect.

Signed-off-by: William Roberts <william.c.roberts@intel.com>