tools: add a flag to display or not usage text
Not all tools require an option in order to be executed correctly, some
tools can be executed without providing any option to them.

Add a boolean so tools can choose whether they want their usage text to
be displayed or not if a user executes them without providing an option.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>