tools: add support to print usage summary when executed with no options
When the tools are executed with no options, print a usage text so users
know what are the available options. This may be useful in systems where
the man pages are not installed.

After this patch the tools will print a usage text like the following:

$ tpm2_activatecredential
usage: tpm2_activatecredential [OPTIONS]
[ -H | --handle=VALUE] [ -c | --context=VALUE] [ -k | --key-handle=VALUE] [ -C | --key-context=VALUE]
[ -P | --Password=VALUE] [ -e | --endorse-passwd=VALUE] [ -f | --in-file=VALUE] [ -o | --out-file=VALUE]
[ -X | --passwdInHex]

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>