tpm2_hash: make hierarchy optional, default to TPM2_RH_OWNER
Make the hierarchy/H argument optional and default to TPM_RH_OWNER when
a value hasn't been specified.

Fix tests/hash: to explicitly pass a null hierarchy now that the default
is owner. With a non-null hierarchy the tpm2_hash command emits a ticket
as well as the hashed value, which confuses the YAML parser.

Signed-off-by: Joshua Lock <joshua.g.lock@intel.com>