session as password
Many tools need to specify N authorizations, where N > 0. Since a
single option cannot satisify this need,  and we want to
limit option explosion, move the session authorizaiotn logic
into the existing password logic. Now, sessions can be passed
via session file, be specifying "session:<path>" as the password.

Passwords have now been revamped to be clear that this is really one
of three authorization mechanisms and their respective options have
changed.

-a/--hierarchy is ALWAYS the hierarchy
??/--auth-XXX always auths to object XXX

The ?? tend to be -P most of the time.

The values chosen for the options might be subject to change in future
changes, but they are at least consistent for now.

Fixes: #904

Signed-off-by: William Roberts <william.c.roberts@intel.com>