options: use TPM2TOOLS_TCTI as environment variable for configuration
The tools used two different environment variables to configure the TCTI:
TPM2TOOLS_TCTI_NAME that was used to set TCTI name and per TCTI variable
that was used to set properties that were specific to a particular TCTI.

For example, the device TCTI used TPM2TOOLS_DEVICE_FILE to choose which
TPM character device is used and so on. Later this was changed to only
need a single variable and both the TCTI name and properties could be
set by separating them using a ':' character.

But the variable name was left as TPM2TOOLS_TCTI_NAME which is kind of
misleading due being formerly use just to configure the TCTI name. So I
think that's better to change the variable name to TPM2TOOLS_TCTI which
is a more generic name and better reflects current variable's semantics.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>