symmetric: rectify symmetric group options
algorithm changes to halg to be in line with non-group options.
Other groups cannot change to algorithm becuase they need both
a key and hash halgorithm in some cases.

The rest of the options, are found to be OK and in line with non-group
tools.

Update the CHANGELOG and analyze script.

Signed-off-by: William Roberts <william.c.roberts@intel.com>