tools: Fix save and load for TPM2B_PRIVATE
The tpm2_create tools is only storing the TPM2B_PRIVATE .buffer field but
not it's .size field. This causes the private keys stored with the stable
version of the tools, to fail loading when using the latest master branch.

Fix this by storing the whole TPM2B_PRIVATE object, including size field.

Fixes: #976

Signed-off-by: William Roberts <william.c.roberts@intel.com>