Fix save and load for TPM2B_PRIVATE
On branch 3.X, the TPM2B_PRIVATE data structure was being serialized to
disk not using libmu and only the buffer is output. On master, this
causes the load with libmu to fail, as it's missing the size UINT16.

Update the load code to handle a TPM2B_PRIVATE saved to disk without
a size field, and issue a warning. Subsequent tools, that write this
to disk will use libmu to correct this.

Fixes: #976

Signed-off-by: William Roberts <william.c.roberts@intel.com>