session: Move session context load from 'save' to 'restore' function.
When a tool wants the session its using to persist across tool
invocation it must save the context to disk so that the next tool that
comes along can reload it using ContextLoad. Previously the
`tpm2_session_save` function was saving and then immediately loading
the session context. The handle was passed to the next tool but the
session had already been flushed by the tabrmd.

To fix this, when the session save function writes the session
structure to disk we now append associated context to the same file
using existing `files_save_tpm_context_to_file` function. In the
`tpm2_session_restore` function we read the tpm2_session structure back,
then read back the context structure and load it into the TPM.

Since we're now loading the context into the TPM when the session is read
back from disk we need access to a sys context. This means we can't load
sessions from the `on_option` callback any longer (no TSS2_SYS_CONTEXT
available at this time). Fixing this required moving all uses of the
`tpm2_auth_util_from_optarg` function (which now takes a
TSS2_SYS_CONTEXT as its first parameter) from `on_option` to the
`tpm2_tool_onrun` function. In the `on_option` callback we now track
whether or not the option was provided and the value provided by the
caller in the tools local context object.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>