autotools: allow man page installation without pandoc being available
With this change the man page install targets will be setup even if
pandoc is not installed but generated man pages are available in
man/man1.

Otherwise without having pandoc available the man pages won't be
installed in case of the release tarball although they are contained in
the tarball.

Signed-off-by: Matthias Gerstner <matthias.gerstner@suse.com>