coverity: fix scanning
Travis' Coverity addon needs access to the build artifacts and doesn't know
how to look into the container, furthermore the Trusty environment in Travis
is difficult to set up with all of the dependencies for the tpm-software
stack (due to conflicting files in packages and requiring newer package
versions).

Therefore instead we run the Coverity tools ourselves instead of relying on
Travis' automation: we ensure the tools are available inside the container,
perform the build with cov-int and bundle up the output for submission
outside of the container.

Signed-off-by: Joshua Lock <joshua.g.lock@intel.com>