tpm2_encryptdecrypt: support iv and algs
Support handling an input and output IV as well as specifying
algorithm mode or figuring it out when not specified.

If the user specifies a mode, use it as is. If the user does not specify a
mode, use the default mode of the object or CFB is the mode of the object
is TPM2_ALG_NULL.

Signed-off-by: William Roberts <william.c.roberts@intel.com>