tpm2_encryptdecrypt: fix mode input
Correct using mode input to tpm2_encryptdecrypt and add tests.

The mode option was added, but unused. Pass the mode to
the TPM for use. The TPM allows the following combinations:
1. If the object has a NULL mode, any valid mode is allowed.
2. If the object has an explicit mode (like cbc) only that mode is allowed.

If the user specifies a mode, that mode is used.
If the user does not specify a mode, either the mode set in object creation
is used, or if NULL mode CFB is used.

Signed-off-by: William Roberts <william.c.roberts@intel.com>