tpm2_import: restore test
tpm2_import test was broken due to changes introduced in PR #1063.

The reason for this, is that the symmetric details of public structure
for the primary object changed from AES128CFB to AES256CFB. It appears

The import tool is hardcoded to a 128 bit AES key for that object.

TODO: remove this hardcode

Fixes: #1079

Signed-off-by: William Roberts <william.c.roberts@intel.com>