tpm2_tool: show usage info when tpm2_tool_onrun returns -1
We want some tools to be able to show usage information under certain
circumstances, such as when a mandatory option was not passed. In order to
facilitate this we now check for one of three return codes from
tpm2_tool_onrun(). As before; 1 indicates an error and 0 indicates success.
Additionally -1 can be returned to indicate to the tool harness that the
usage information should be shown.

Signed-off-by: Joshua Lock <joshua.g.lock@intel.com>