tpm2_getcap: restore tool output to print properties with TPM_PT prefix
Commit ac9a5d787753 ("update to TSS version 2.0") updated tpm2-tss version
to 2.0. On tpm2-tss, the TPM_PT_* constants where renamed to TPM2_PT_* but
that commit also changed the output of the tpm2_getcap tool to match the
name of the properties.

Unfortunately this is a bacward incompatible change that could break users
that were relying on the tpm2_getcap tool output.

Restore the older output and also fix the tests that check these values.

Fixes: #1175

Reported-by: Vilem Marsik <vmarsik@redhat.com>
Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>