tpm2_auth_util: cleanup session auth value parsing
Rather than have the general entry point deal with the parsing
pf the +, which would break passwords with a + in them, check
if it is a session, copy and modify the argument spliting the
string on the + character and call into the password handler for
the remaining password string if present.

This will support passwords with strings, and remove some
complexity from the code (confusing loop).

Signed-off-by: William Roberts <william.c.roberts@intel.com>