lib/tpm2_util: enhance tpm2_util_object_load to return the objects type
It may be useful in future for various tools to know whether the loaded
object was a handle id (hex or scalar) or whether it was a context file.
To that end have the helper function tpm2_util_object_load() return a
custom type indicating whether the function failed to load an object
(olrc_error), whether the loaded object was a TPM2_HANDLE (olrc_handle) or
whether the loaded object was a file (olrc_file).

Update all callers of the function to handle the new return value.

Signed-off-by: Joshua Lock <joshua.g.lock@intel.com>