ci: Dockerfile of our container for CI
We need to update our CI container images to include a newer OpenSSL, as the
version in Ubuntu 16.04 has memory leaks exposed by the ESAPI TSS backend which
ultimately result in CI failures for our clang build with the ESAPI port.

As a first step on this path attempt to replicate the current hand-stitched
container image with a Dockerfile.

Signed-off-by: Joshua Lock <joshua.g.lock@intel.com>