build: move to recommended bash-completion directory
/etc/bash_completion.d (compatdir) is deprecated in favour of
/usr/share/bash-completion/completions (completionsdir), see the bash-completion FAQ.
As the completions in the new directory are loaded on demand based on their
filename, we need to create symlinks for all tools during installation.

The new on demand loading allows to replace the find-based loop to initialise
completion for all tools with an approach based on the currently requested completion.