tpm2_changeauth: fix transient object changeauth
Populate the ESYS_TR handle when given a TPM2_HANDLE. When specifying an object
via a TPM2_HANDLE provided by -a, convert that handle into an ESYS_TR handle
identifier.

Additionally, the switch statement verifying the type of handle, PCR, transient, etc
cannot check the TPM2_HANDLE on the transient case, as the handle is 0 and is not
exposed via the ESYS_TR paradigm.

This fixes transient object change auth.

Signed-off-by: William Roberts <william.c.roberts@intel.com>