build: Use dbus-run-session in automake test harness.
This commit removes the manual launching of the dbus session bus
instance in the docker.run file. This is replaced by adding the
'dbus-run-session' command to the AM_TEST_ENVIRONMENT. This allows
callers to simply execute 'make check' and the test harness will set up
a session bus instance automagically.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>