configure.ac: tweak output of ./configure --help
The users asume, unless otherwise stated, that a feature has two states
and it is either disabled or enabled.  For features enabled by default,
the output of “./configure --help” is supposed to state how ot disable them
and for features disabled by default, the very same output is supposed to
state how to enable them.

This change adjusts the output of “./configure --help” accordingly.

Signed-Off-By: Дилян Палаузов <git-dpa@aegee.org>