MAINTAINERS: Remove myself
I've been busy with other projects lately and couldn't find time to do the
work expected for someone listed as a maintainer of the project. So remove
myself from the MAINTAINERS.md file.

I will continue maintaining the tpm2-{tools,tss,abrmd} packages in Fedora
and try to help with this project but let's make the file reflect reality.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>