lib: Remove hardcoded boundaries on pcrs list/banks
The struct "tpm2_algorithm" is an intermediate layer that
represent the number of banks available and supported
by TPM we are about to communicate with. It seems
logical to support up to TPM2_NUM_PCR_BANKS to
conform with TSS and the possibility of the TPM having
more than the common 24 PCRs

Altough the "tpm2_pcrs" is the same with PCR lists,
this is somehow different as we are just limiting the
number of PCR the tool is able to dump at once.

For the sake of consistency with what was the
original behaviour, just allow the user to
dump at max a complete bank.

Signed-off-by: sebastien le stum <lestums@gmail.com>