test/integration/tests: Fix pcrextend test suite on overlength
Test was correct but not testing what the comment stated : As $sha1hash
variable was never declared, the test was actually testing the length of
alg being inferior of the expected size for specified algorithm. However
this test was supposed to test an oversized spec.

Fixed the test and added a test on wrong sized hex string for hash (sha256
value for a sha1 bank)

Also removed the disabling of the TRAP

Signed-off-by: sebastien le stum <lestums@gmail.com>