lib/tpm2_options: remove -X option
The -X option was introduced as a global option, which presented some
issues, notably:
1. Possible NULL pointer dereferences on tools unaware of a possible
   NULL ESAPI.
2. Globaly reserved option for many tools that don't need it.

To correct this, make a new ESAPI flag for optional SAPI initialization
and couple it with a new TCTI called "none". Tools that have this
optional SAPI flag set, will also support --tcti=none. Tools that don't
support the "none" tcti, will throw an error. Update the makecredential
test which was using -X and and a negative test to getrandom to make
sure the none tcti fails properly when a tool requires a SAPI.

Fixes: #1442

Signed-off-by: William Roberts <william.c.roberts@intel.com>