Convert to SPDX-License-Identifiers - BSD-3-Clause
Instead of adding the same 'boilerplate' license text we can use the much
more concise SPDX Short License IDentifiers.

The patch replaces the standard BSD-3 Clause License text with
SPDX-License-Identifier: BSD-3-Clause.
It does not change the license of the file itself - no other changes.

Only exception is that in some shell scripts the shebang was moved
to the top of the file.

Signed-off-by: Peter Huewe <peter.huewe@infineon.com>