ci: check for whitespace errors against $TRAVIS_BRANCH
Currently the whitespace error check is always performed against the
master branch. Therefore it fails on different branches because Travis
CI only does a shallow clone of the current branch. Replacing it with
$TRAVIS_BRANCH allows to test against the correct base branch for pull
requests and is a no-op if a branch is built directly. In case the CI
scripts are used locally without Travis, use the master branch as a
default for $TRAVIS_BRANCH to retain the current behaviour.

Signed-off-by: Jonas Witschel <diabonas@gmx.de>