tpm2_tcti_ldr: safe print option string
The TCTI option string is allowed to remain NULL, and thus
passing it as an argument to printf("%s") is not portable.
GLIBC will print (null) but some systems will crash.

Define a macro that will return the string "(null)" when the
string is NULL so other LIBC's have the same behavior as
GLIBC.

Fixes: #1241

Signed-off-by: William Roberts <william.c.roberts@intel.com>