tpm2_getcap: replace value to raw when not modified
In most cases, the value in tpm2_getcap -c properties-fixed
were output in a raw hex format and not shown modified in any
way. The raw field should be used, so if value ever needs to
be added, we can just add in that field without needing to
change it.

Signed-off-by: William Roberts <william.c.roberts@intel.com>