tests: have bash function ina() print the result instead of returning it
The bash return code is actually an exit status which indicates
success/error, not the result of the logic of the function. Hence, the
bash function ina() in helpers.sh should print its result rather than
returning it. This way, the bash ERR trap does not falsely recognize the
function's result as an error.

Signed-off-by: joholl <johannes.holland@infineon.com>