tpm2_import: fix return value on file context load
In files_load_tpm_context_from_path() the rc was still being returned as
if files_load_tpm_context_from_file() returned a bool. However, it
returns an rc, and thus the success of 0 was causing the rc to be
coerced to a failure. Just return the rc as is.

Signed-off-by: William Roberts <william.c.roberts@intel.com>