tests: interpret every key/value as string in yaml_get_kv()
The function yaml_get_kv interprets hashes as numerical values and prints
them in decimal. As a consequence, comparing a hash to an expected value
fails in loadexternal.sh. Additionally, parsing numerical values is an
unused feature and makes parsing strings more complicated. As a
consequence, everything should be parsed as string.

Signed-off-by: joholl <johannes.holland@infineon.com>