tpm2_activatecredential: Update hardcoded auth interface of the credential key to generic
The keyhandle authorization in the activatecredential tool should generically
take auth in form of a password, pcr, session, etc. Thus, following any possible
way of providing auths in the cmdauths structure. Currently activatecredential
is hardcoded to satisfy a policysecret of the endorsement hierarchy. This commit
fixes this issue by decoupling it from the hardcoded auth method. Essentially
this means after this commit, if there is need for activating credentials for
the activate handle where in the keyhandle is the EK whose auth policy is setup
per TCG prescribed auth policy for EK; then that auth policy ought to be
satisfied which is the policysecret of the endorsement hierarchy auth. This
satisfied policy session is then used as the auth using the common auth
interface in the tool.

Signed-off-by: Imran Desai <imranodesai@gmail.com>