test/integration: check if port is bound to the simulator
On some systems problems seem to occur when running many tests in
parallel because of a race condition between checking whether the
simulator ports are free and starting the simulator on the specified
ports. Fix this by trying to start the simulator first, then check
whether it is listening on the ports and retry if necessary.

Signed-off-by: Jonas Witschel <diabonas@gmx.de>