test: control setting which TCTI tests run to
The tests before made some asumptions about the environment in which
they were started, allow finer grained control of testing environment
for those wishing to use the test harness out of the automake
environment. This can be configured through:

TPM2_SIM: the simulator executable, or none to not use a simulator.
TPM2_SIMPORT: the port to use for the simulator, or none to autopick.

TPM2_ABRMD: the abrmd executable, or none to not use an RM.
TPM2ABRMD_TCTI: the abrmd tcti to use, or none to use simulator port.

TPM2TOOLS_TEST_TCTI: is specified it is the TCTI for the tools to use,
  or autopick based on:
    - if abrmd was started use it.
    - if abrmd was not started use simulator.
    - if neither was specified it's an error. Do not default to device,
      we don't generally want to run test scripts at devices. By
      explicitly setting TPM2TOOLS_TCTI you can speciffy device.

We explicitly use the variable TPM2TOOLS_TEST_TCTI to be distinct from
the env variable used in the tools themselved, TPM2TOOLS_TCTI, to
prevent having a device TCTI configured in the environment and calling
make check and having it attempt to run potentially dangerous scripts
against the physical TPM.

Example running a specific test against a TPM device:
  - make check TPM2TOOLS_TEST_TCTI=device:/dev/tpm0 TESTS=test/integration/tests/checkquote.sh

    ERROR:tcti:src/tss2-tcti/tcti-device.c:399:Tss2_Tcti_Device_Init() Failed to open device file /dev/tpm0: Permission denied
    ERROR: tcti init allocation routine failed for library: "device" options: "/dev/tpm0"
    ERROR: Could not load tcti, got: "device"

  optionally you can unset TPM2_ABRMD and TPM2_SIMLATOR to prevent those
  from starting up using automake TESTS_ENVIRONMENT variable.

  make check TESTS_ENVIRONMENT='TPM2_ABRMD="" TPM2_SIM=""' TPM2TOOLS_TEST_TCTI=device:/dev/tpm0 TESTS=test/integration/tests/checkquote.sh

Example running a specific test against the SIMULATOR no ABRMD:
    TPM2TOOLS_TEST_TCTI not set, attempting to figure out default
    export TPM2TOOLS_TCTI="mssim:host=localhost,port=5055"

Example running with ABRMD connected to a device:
  - make check TESTS_ENVIRONMENT='TPM2_SIM=""' TPM2ABRMD_TCTI=device:/dev/tpm0 TESTS=test/integration/tests/checkquote.sh

Signed-off-by: William Roberts <william.c.roberts@intel.com>