tpm2_auth_util: add interactive password support
Add support for file:- where it can figure out if it's a TTY
and prompt the user for the password disabling echoing back.

Also, piping and here strings tend to get a trailing newline
so remove them.

Use getline() when reading the userinput from TTY so an enter
will end the password prompting portion and strip the resulting
newline.

Signed-off-by: Imran Desai <imranodesai@gmail.com>
Signed-off-by: William Roberts <william.c.roberts@intel.com>