tpm2_nvread.c: Specify handle flags to filter allowed handle types
Specify handle flags for filtering

Prior to the commit that added filtering for NV handles; the handle
filter was setup with flags to accept all TPM2 handles. This commit
makes it explicit through flags that the handle type has to be of type
owner hierarchy or platform hierarchy or a valid NV index.

Signed-off-by: Imran Desai <imran.desai@intel.com>