tpm2_load: output name
The manpage and ticket #1042 both discuss the name output from
the tpm2_load command. However, this was missing. Output the name
as a YAML key value pair with the key of name. This way the output
is both clear in what is being output and also in the
future we can add additional outputs if needed.

Signed-off-by: William Roberts <william.c.roberts@intel.com>