lib/tpm2.c: Fix issue with status return from encryptdecrypt call
There was a bug in the tpm2_encryptdecrypt routine where in if the
EncryptDecrypt2 command failed the rval was wrongly compared to
TPM2_RC_COMMAND_CODE instead of TPM2_RC_SUCCESS. This caused the
command to return successfully even in the event of a failure.
Thus causing segmentation faults in the calling program which
relied on this error code to skip freeing buffers.

Signed-off-by: Imran Desai <imran.desai@intel.com>