test: remove external hex conversion of random bytes
For quoting, 20 random bytes are needed. However, the casting of these
random bytes into a hex stream is done two times, first with --hex
(introduced with 460afc31) and then by calling xxd. This results in
80 characters (representing 40 bytes) which is not what we want.
With the internal conversion of tpm2_getrandom into a hex stream, the call
of xxd can be ommitted.

This is a fixup to patch a25dff.

Signed-off-by: William Roberts <william.c.roberts@intel.com>