tcti: use separate rc variable for Tss2_TctiLdr_*
The return code of Tss2_TctiLdr_* is of type TSS2_RC, while tpm2-tools
internally uses tpm2_option_code. This results in a zero exit status
even if the TCTI could not be initialised because the return code is
not interpreted correctly. Also add a unit test for this scenario.

Fixes: GH-1617
Fixes: 0505f2b3969e8ca2006dcfbb083d45a3b62bfb30
Signed-off-by: Jonas Witschel <diabonas@gmx.de>