tcti: allow specifying long option --tcti without equals sign
If 'optional_argument' is used instead of 'required_argument', only the
syntax '--tcti=<value>' is accepted, but not '--tcti <value>'. Since
this does not match the definition of the short option as required in
'common_short_opts' and using --tcti without an argument does not make
sense anyway, we use required_argument and add a unit test for the case
without equals sign.

Fixes: 56710b6e1c4d3e02ff77ad367c216a59b51ca765
Signed-off-by: Jonas Witschel <diabonas@gmx.de>