lib/tpm2_auth_util: change +file to =file in PRC mini language
A plus is already used as the separator between different banks, hence
a specification with multiple banks like

sha256:0,1,2,3+sha1:0,1,2,3

is currently not parsed correctly because the second bank is interpreted
as the PCR file. The equals sign was chosen as the new separator to have
a syntax somewhat similar to tpm2_pcrextend.

Fixes: GH-1635
Signed-off-by: Jonas Witschel <diabonas@gmx.de>