test/integration: test PCR mini language with multiple banks
Previously the PCR mini language specification could only handle one
bank and the feature to specify the expected PCR values was broken when
using more than one bank. This has been fixed in the previous two
commits, so make sure to give this some test coverage in the future by
testing the PCR mini language with SHA-1 and SHA-256 simultaneously.

Signed-off-by: Jonas Witschel <diabonas@gmx.de>