test: add internet connection check for tpm2_getekcertificate.sh
If there is no internet connection, tpm2_getekcertificate.sh will always
fail. With this fix, if curl is installed, a sanity check is performed:
if there is no internet connection or if ekop.intel.com is down,
77 (SKIP) is returned.
If curl is not installed, no sanity check is performed and the
test will fail if the website cannot be reached.

Signed-off-by: joholl <johannes.holland@infineon.com>