lib/pcr: fix bounds check
The check in the PCR code checks that the number or count
of assigned PCRs is less than the size in bytes of of the
tpm2_algorithm structure. We really want to make sure
that the count of source is less than or equal the
count of destination. Correct this by using ARRAY_LEN() vs
sizeof().

Signed-off-by: William Roberts <william.c.roberts@intel.com>