tpm2_send: a tainted buffer can cause memory overun
If the input buffer is malformed so that the command size is smaller
than the data size, the malloc of command header will be to small to
accommodate teh fread() of the data buffer. Ensure that the buffer
will be big enough by checking that the overall command size is greater
than its payload size.

Signed-off-by: William Roberts <william.c.roberts@intel.com>