tpm2_cc_util: remove error message
Sometime shwen converting a command to a name, the command isn't found.
That's ok, some manufacturers have custom commands or the TPM is on a
newer spec then the tools. The tpm2_cc_util routine for this conversion
already handled the not found case, and just returned NULL. The caller
in tpm2_getcap would then create a name for this tool. However, the
conversion routine was printing an error message that really wasn't on
a true error case, so remove it.

Fixes: #1704

Signed-off-by: William Roberts <william.c.roberts@intel.com>